Add the python integration test to workflow
Now that the test is working we should add it to the workflow. I have
left this as a seperate commit in case it adds too much time to the
workflows.

By adding ubuntu latest for linux and a special workflow command for
macos we are able to enable redis to create a compatible workflow for
running the new integration tests.